+++ src/main/org/firebirdsql/gds/impl/GDSServerVersion.java
@@ -1,30 +1,31 @@
 /*
  * Public Firebird Java API.
  *
- * Redistribution and use in source and binary forms, with or without 
+ * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *    1. Redistributions of source code must retain the above copyright notice, 
+ *    1. Redistributions of source code must retain the above copyright notice,
  *       this list of conditions and the following disclaimer.
- *    2. Redistributions in binary form must reproduce the above copyright 
- *       notice, this list of conditions and the following disclaimer in the 
- *       documentation and/or other materials provided with the distribution. 
- *    3. The name of the author may not be used to endorse or promote products 
+ *    2. Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *    3. The name of the author may not be used to endorse or promote products
  *       derived from this software without specific prior written permission.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE AUTHOR ``AS IS'' AND ANY EXPRESS OR IMPLIED 
- * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF 
- * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO 
- * EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, 
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, 
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; 
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR 
- * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF 
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR ``AS IS'' AND ANY EXPRESS OR IMPLIED
+ * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
+ * EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+ * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 package org.firebirdsql.gds.impl;
 
 import java.io.Serializable;
+import java.util.Arrays;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -33,7 +34,8 @@ import java.util.regex.Pattern;
  * in response to the <code>isc_info_firebird_version</code> information call.
  * Expected version format is:
  * <p>
- * <code>&lt;platform&gt;-&lt;type&gt;&lt;majorVersion&gt;.&lt;minorVersion&gt;.&lt;variant&gt;.&lt;buildNum&gt; &lt;serverName&gt;[,&lt;extended server info&gt;]</code>
+ * <code>&lt;platform&gt;-&lt;type&gt;&lt;majorVersion&gt;.&lt;minorVersion&gt;.&lt;variant&gt;.&lt;buildNum&gt;
+ * &lt;serverName&gt;</code>, and additional version string elements if present.
  * </p>
  * <p>
  * where <code>platform</code> is a two-character platform identification string,
@@ -55,15 +57,17 @@ public final class GDSServerVersion implements Serializable {
     public static final String CONNECTION_OPTION_COMPRESSION = "Z";
 
     /**
-     * GDSServerVersion that can be used as a dummy/invalid object when a version object is required, but none is available.
+     * GDSServerVersion that can be used as a dummy/invalid object when a version object is required, but none is
+     * available.
      */
-    public static final GDSServerVersion INVALID_VERSION = new GDSServerVersion("INVALID", "", "", "", 0, 0, 0, 0, "", "");
+    public static final GDSServerVersion INVALID_VERSION =
+            new GDSServerVersion(new String[] { "INVALID" }, "", "", "", 0, 0, 0, 0, "");
 
-    private static final Pattern VERSION_PATTERN = 
-        Pattern.compile("((\\w{2})-(\\w)(\\d+)\\.(\\d+)\\.(\\d+)\\.(\\d+)(?:-\\S+)?) ([^-,/]+)(?:[-,/](.*))?");
+    private static final Pattern VERSION_PATTERN =
+            Pattern.compile("((\\w{2})-(\\w)(\\d+)\\.(\\d+)\\.(\\d+)\\.(\\d+)(?:-\\S+)?) (.+)");
 
     private static final Pattern CONNECTION_METADATA_PATTERN = Pattern.compile("/P(\\d+)(?::([^:]+))?$");
-    
+
     private static final int FULL_VERSION_IDX = 1;
     private static final int PLATFORM_IDX = 2;
     private static final int TYPE_IDX = 3;
@@ -72,10 +76,9 @@ public final class GDSServerVersion implements Serializable {
     private static final int VARIANT_IDX = 6;
     private static final int BUILD_IDX = 7;
     private static final int SERVER_NAME_IDX = 8;
-    private static final int EXTENDED_INFO_IDX = 9;
 
-    private final String rawStr;
-    
+    private final String[] rawVersions;
+
     private final String platform;
     private final String type;
 
@@ -86,11 +89,10 @@ public final class GDSServerVersion implements Serializable {
     private final int buildNumber;
 
     private final String serverName;
-    private final String extendedServerName;
 
-    private GDSServerVersion(String rawStr, String platform, String type, String fullVersion, int majorVersion,
-            int minorVersion, int variant, int buildNumber, String serverName, String extendedServerName) {
-        this.rawStr = rawStr;
+    private GDSServerVersion(String[] rawVersions, String platform, String type, String fullVersion, int majorVersion,
+            int minorVersion, int variant, int buildNumber, String serverName) {
+        this.rawVersions = rawVersions.clone();
         this.platform = platform;
         this.type = type;
         this.fullVersion = fullVersion;
@@ -99,7 +101,6 @@ public final class GDSServerVersion implements Serializable {
         this.variant = variant;
         this.buildNumber = buildNumber;
         this.serverName = serverName;
-        this.extendedServerName = extendedServerName;
     }
 
     public int getBuildNumber() {
@@ -131,28 +132,43 @@ public final class GDSServerVersion implements Serializable {
     }
 
     public String getExtendedServerName() {
-        return extendedServerName;
+        if (rawVersions.length < 2) {
+            return null;
+        } else if (rawVersions.length == 2) {
+            return rawVersions[1];
+        } else {
+            StringBuilder sb = new StringBuilder();
+            for (int idx = 1; idx < rawVersions.length; idx++) {
+                if (idx > 1) {
+                    sb.append(',');
+                }
+                sb.append(rawVersions[idx]);
+            }
+            return sb.toString();
+        }
     }
-    
+
     public String getFullVersion() {
-    	return fullVersion;
+        return fullVersion;
     }
 
     /**
      * @return Protocol version of the connection, or {@code -1} if this information is not available.
      */
     public int getProtocolVersion() {
-        if (extendedServerName == null) return -1;
-        Matcher connectionMetadataMatcher = CONNECTION_METADATA_PATTERN.matcher(extendedServerName);
-        if (connectionMetadataMatcher.find()) {
-            String protocolVersion = connectionMetadataMatcher.group(1);
-            return Integer.parseInt(protocolVersion);
-        }
-        return -1;
+        // We assume the protocol information is in the second version string;
+        // TODO this assumption may be wrong for multi-hop connections
+        if (rawVersions.length == 1 || rawVersions[1] == null) return -1;
+        Matcher connectionMetadataMatcher = CONNECTION_METADATA_PATTERN.matcher(rawVersions[1]);
+        if (!connectionMetadataMatcher.find()) return -1;
+
+        String protocolVersion = connectionMetadataMatcher.group(1);
+        return Integer.parseInt(protocolVersion);
     }
 
     /**
-     * @return {@code true} if encryption is used, {@code false} if no encryption is used <b>or</b> if this information is not available
+     * @return {@code true} if encryption is used, {@code false} if no encryption is used <b>or</b> if this information
+     * is not available
      */
     public boolean isWireEncryptionUsed() {
         return getConnectionOptions().contains(CONNECTION_OPTION_ENCRYPTED);
@@ -163,8 +179,10 @@ public final class GDSServerVersion implements Serializable {
     }
 
     private String getConnectionOptions() {
-        if (extendedServerName == null) return "";
-        Matcher connectionMetadataMatcher = CONNECTION_METADATA_PATTERN.matcher(extendedServerName);
+        // We assume the protocol information is in the second version string;
+        // TODO this assumption may be wrong for multi-hop connections
+        if (rawVersions.length == 1 || rawVersions[1] == null) return "";
+        Matcher connectionMetadataMatcher = CONNECTION_METADATA_PATTERN.matcher(rawVersions[1]);
         if (!connectionMetadataMatcher.find()) return "";
 
         String connectionOptions = connectionMetadataMatcher.group(2);
@@ -172,7 +190,7 @@ public final class GDSServerVersion implements Serializable {
     }
 
     public int hashCode() {
-        return rawStr.hashCode();
+        return Arrays.hashCode(rawVersions);
     }
 
     public boolean equals(Object obj) {
@@ -182,46 +200,62 @@ public final class GDSServerVersion implements Serializable {
 
         GDSServerVersion that = (GDSServerVersion) obj;
 
-        return rawStr.equals(that.rawStr);
+        return Arrays.equals(this.rawVersions, that.rawVersions);
     }
 
     public String toString() {
-        return rawStr;
+        if (rawVersions.length == 1) {
+            return rawVersions[0];
+        }
+        StringBuilder sb = new StringBuilder();
+        int idx = 0;
+        sb.append(rawVersions[idx++]);
+        do {
+            sb.append(',');
+            sb.append(rawVersions[idx++]);
+        } while (idx < rawVersions.length);
+        return sb.toString();
     }
 
     /**
      * Parse the raw version string and create a GDSServerVersion object.
-     * 
-     * @param versionString string to parse.
-     * 
-     * @throws GDSServerVersionException if versionString does not match expected pattern
+     *
+     * @param versionStrings
+     *         strings to parse, expects a non-empty array with at least 1, and usually 2 version strings
+     * @throws GDSServerVersionException
+     *         if versionString does not match expected pattern
      */
-    public static GDSServerVersion parseRawVersion(String versionString) throws GDSServerVersionException {
-        Matcher matcher = VERSION_PATTERN.matcher(versionString);
+    public static GDSServerVersion parseRawVersion(String... versionStrings) throws GDSServerVersionException {
+        if (versionStrings == null || versionStrings.length == 0 || versionStrings[0] == null) {
+            throw new GDSServerVersionException("No version string information present");
+        }
+
+        Matcher matcher = VERSION_PATTERN.matcher(versionStrings[0]);
         if (!matcher.matches()) {
-            throw new GDSServerVersionException(String.format("Version string \"%s\" does not match expected format",
-            		versionString));
+            throw new GDSServerVersionException(
+                    String.format("Version string \"%s\" does not match expected format", versionStrings[0]));
         }
 
         return new GDSServerVersion(
-                versionString, 
+                versionStrings,
                 matcher.group(PLATFORM_IDX),
-                matcher.group(TYPE_IDX), 
-                matcher.group(FULL_VERSION_IDX), 
+                matcher.group(TYPE_IDX),
+                matcher.group(FULL_VERSION_IDX),
                 Integer.parseInt(matcher.group(MAJOR_IDX)),
-                Integer.parseInt(matcher.group(MINOR_IDX)), 
+                Integer.parseInt(matcher.group(MINOR_IDX)),
                 Integer.parseInt(matcher.group(VARIANT_IDX)),
-                Integer.parseInt(matcher.group(BUILD_IDX)), 
-                matcher.group(SERVER_NAME_IDX),
-                matcher.group(EXTENDED_INFO_IDX));
+                Integer.parseInt(matcher.group(BUILD_IDX)),
+                matcher.group(SERVER_NAME_IDX));
     }
 
     /**
      * Convenience method to check if the major.minor of this version is equal to or larger than the specified required
      * version.
      *
-     * @param requiredMajorVersion Required major version
-     * @param requiredMinorVersion Required minor version
+     * @param requiredMajorVersion
+     *         Required major version
+     * @param requiredMinorVersion
+     *         Required minor version
      * @return <code>true</code> when current major is larger than required, or major is same and minor is equal to or
      * larger than required
      */
@@ -229,4 +263,5 @@ public final class GDSServerVersion implements Serializable {
         return majorVersion > requiredMajorVersion ||
                 (majorVersion == requiredMajorVersion && minorVersion >= requiredMinorVersion);
     }
+    
 }